Move RTE components to a separate dir
